pyproject.toml
[project]
name = "signapp"
version = "0.1.0"
description = "Audio processing pipeline for sign language recognition with Whisper transcription and disfluency removal"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "SignApp Contributors"}
]
keywords = ["audio", "transcription", "whisper", "disfluency", "sign-language", "nlp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "openai-whisper>=20231117",
    "torch>=2.0.0",
    "transformers>=4.34.0",
    "datasets>=2.14.0",
    "mlflow>=2.10.0",
    "evaluate>=0.4.0",
    "nltk>=3.8.1",
    "requests>=2.31.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.10.0",
    "ruff>=0.14.10",
    "mypy>=1.6.0",
]

cuda = [
    "torch[cuda]>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/SignApp"
Documentation = "https://github.com/yourusername/SignApp/blob/main/README.md"
Repository = "https://github.com/yourusername/SignApp.git"
"Bug Tracker" = "https://github.com/yourusername/SignApp/issues"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
